Let the user handle the real size of SHM segments.
Some OSes (e.g. Windows, OSX) allocate shared memory in multiple of page
sizes. We don't store the real size in the memory segment any more and
let the users do that themselves if necessary. For this purpose the new
`h:truncate()` method has been added to the memfile interface.